Removed unused functions
